BF: Don't add --user flag if --target is specified

+++ psychopy/tools/pkgtools.py
@@ -246,26 +246,33 @@ def installPackage(
         If `awaited=False`:
             Returns the job (thread) which is running the install.
     """
-    if target is None:
-        target = prefs.paths['userPackages']
-
     # convert extra to dict
     if extra is None:
         extra = {}
-    # check the directory exists before installing
-    if not os.path.exists(target):
-        raise NotADirectoryError(
-            'Cannot install package "{}" to "{}", directory does not '
-            'exist.'.format(package, target))
 
     # construct the pip command and execute as a subprocess
     cmd = [sys.executable, "-m", "pip", "install", package]
 
     # optional args
     if target is None:  # default to user packages dir
-        cmd.append('--prefix')
-        cmd.append(prefs.paths['packages'])
+        # check if we are in a virtual environment, if so, dont use --user
+        if hasattr(sys, 'real_prefix') or (
+                hasattr(sys, 'base_prefix') and sys.base_prefix != sys.prefix):
+            # we are in a venv
+            logging.warning(
+                "You are installing a package inside a virtual environment. "
+                "The package will be installed in the user site-packages "
+                "directory."
+            )
+        else:
+            cmd.append('--user')
     else:
+        # check the directory exists before installing
+        if target is not None and not os.path.exists(target):
+            raise NotADirectoryError(
+                'Cannot install package "{}" to "{}", directory does not '
+                'exist.'.format(package, target))
+
         cmd.append('--target')
         cmd.append(target)
     if upgrade:
@@ -274,17 +281,6 @@ def installPackage(
         cmd.append('--force-reinstall')
     if noDeps:
         cmd.append('--no-deps')
-
-    # check if we are in a virtual environment, if so, dont use --user
-    if hasattr(sys, 'real_prefix') or (
-            hasattr(sys, 'base_prefix') and sys.base_prefix != sys.prefix):
-        # we are in a venv
-        logging.warning(
-            "You are installing a package inside a virtual environment. "
-            "The package will be installed in the user site-packages directory."
-        )
-    else:
-        cmd.append('--user')
 
     cmd.append('--prefer-binary')  # use binary wheels if available
     cmd.append('--no-input')  # do not prompt, we cannot accept input
